requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
black==23.11.0
bleach==6.1.0
blinker==1.7.0
bs4==0.0.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.20
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.0
dataclasses-json==0.6.3
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.8.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.105.0
fastjsonschema==2.19.0
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==23.5.26
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.12.2
future==0.18.3
google-auth==2.25.2
googleapis-common-protos==1.62.0
greenlet==3.0.1
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.6
importlib-metadata==6.8.0
importlib-resources==6.1.1
iniconfig==2.0.0
ipykernel==6.27.1
ipython==8.18.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.11.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kubernetes==28.1.0
langchain==0.0.344
langchain-core==0.0.8
langsmith==0.0.67
llvmlite==0.41.1
loguru==0.7.2
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mistune==3.0.2
mmh3==4.0.1
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numpy==1.26.2
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.3.6
openai-whisper==20231117
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-asgi==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
overrides==7.4.0
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
platformdirs==4.0.0
pluggy==1.3.0
posthog==3.1.0
prettytable==3.9.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.6
pulsar-client==3.3.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.17.2
PyPika==0.48.9
pyreadline3==3.4.1
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.2
rsa==4.9
safetensors==0.4.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
terminado==0.18.0
tiktoken==0.5.2
tinycss2==1.2.1
tokenize-rt==5.2.0
tokenizers==0.15.0
tomli==2.0.1
torch==2.1.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
transformers==4.36.2
typer==0.9.0
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.18
uvicorn==0.24.0.post1
watchfiles==0.21.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==12.0
Werkzeug==3.0.1
win32-setctime==1.1.0
wrapt==1.16.0
yarl==1.9.3
zipp==3.17.0
